requirements.txt
numpy~=1.22.3
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
opencv-python~=4.5.5.64
Pillow~=9.1.1